use the same knownPackage definition to avoid drift between the two places which use the definion
